requirements.txt
streamlit>=1.28.0
plotly>=5.15.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
ta>=0.10.0
python-dotenv>=0.19.0